# apps/example-nest/src/orders.controller.ts
import { Body, Controller, Delete, Get, Headers, Param, Patch, Post, Query } from '@nestjs/common';
import {
  ApiBearerAuth,
  ApiBody,
  ApiCreatedResponse,
  ApiExtraModels,
  ApiHeader,
  ApiNoContentResponse,
  ApiOkResponse,
  ApiOperation,
  ApiQuery,
  ApiTags,
  getSchemaPath,
} from '@nestjs/swagger';
import { Address, CreateOrderBody, Order, OrderLineItem, ShipmentTrackingEventNotificationDelivery } from './orders.dto.js';

@ApiTags('orders')
@ApiBearerAuth()
@Controller('orders')
@ApiExtraModels(Order, OrderLineItem, Address, CreateOrderBody, ShipmentTrackingEventNotificationDelivery)
export class OrdersController {
  @Get()
  @ApiOperation({ summary: 'List orders' })
  @ApiQuery({ name: 'status', required: false, enum: ['pending', 'processing', 'shipped', 'delivered', 'cancelled', 'refunded'] })
  @ApiQuery({ name: 'customerId', required: false })
  @ApiQuery({ name: 'from', required: false, schema: { type: 'string', format: 'date' } })
  @ApiQuery({ name: 'to', required: false, schema: { type: 'string', format: 'date' } })
  @ApiQuery({ name: 'page', required: false, schema: { type: 'integer', default: 1 } })
  @ApiQuery({ name: 'limit', required: false, schema: { type: 'integer', default: 20 } })
  @ApiHeader({ name: 'x-store-id', required: true })
  @ApiOkResponse({
    description: 'Paginated orders list.',
    schema: {
      type: 'object',
      properties: {
        data: { type: 'array', items: { $ref: getSchemaPath(Order) } },
        total: { type: 'integer' },
        page: { type: 'integer' },
      },
    },
  })
  listOrders(@Headers('x-store-id') storeId?: string) {
    return { data: [], total: 0, page: 1 };
  }

  @Post()
  @ApiOperation({ summary: 'Create a new order' })
  @ApiHeader({ name: 'x-idempotency-key', required: true })
  @ApiHeader({ name: 'x-store-id', required: true })
  @ApiBody({ schema: { $ref: getSchemaPath(CreateOrderBody) } })
  @ApiCreatedResponse({ description: 'Order created.' })
  createOrder(@Body() body: any) {
    return { id: 'ord_1', status: 'pending', ...body };
  }

  @Get('stats')
  @ApiOperation({ summary: 'Get order statistics' })
  @ApiQuery({ name: 'period', required: false, enum: ['today', 'week', 'month', 'quarter', 'year'], schema: { default: 'month' } })
  @ApiHeader({ name: 'x-store-id', required: true })
  @ApiOkResponse({
    description: 'Order statistics.',
    schema: {
      type: 'object',
      properties: {
        totalOrders: { type: 'integer' },
        totalRevenue: { type: 'number' },
        averageOrderValue: { type: 'number' },
        byStatus: {
          type: 'object',
          additionalProperties: { type: 'integer' },
        },
        topProducts: {
          type: 'array',
          items: {
            type: 'object',
            properties: { productId: { type: 'string' }, name: { type: 'string' }, quantity: { type: 'integer' } },
          },
        },
      },
    },
  })
  getOrderStats() {
    return { totalOrders: 0, totalRevenue: 0, averageOrderValue: 0, byStatus: {}, topProducts: [] };
  }

  @Post('batch')
  @ApiOperation({ summary: 'Perform a batch action on orders' })
  @ApiBody({
    schema: {
      type: 'object',
      required: ['orderIds', 'action'],
      properties: {
        orderIds: { type: 'array', items: { type: 'string' } },
        action: { type: 'string', enum: ['cancel', 'archive', 'export'] },
        reason: { type: 'string' },
      },
    },
  })
  @ApiOkResponse({
    description: 'Batch action result.',
    schema: {
      type: 'object',
      properties: {
        processed: { type: 'integer' },
        failed: { type: 'integer' },
        errors: { type: 'array', items: { type: 'object', properties: { orderId: { type: 'string' }, message: { type: 'string' } } } },
      },
    },
  })
  batchOrders(@Body() body: any) {
    return { processed: 0, failed: 0, errors: [] };
  }

  @Get(':id')
  @ApiOperation({ summary: 'Get order details' })
  @ApiOkResponse({ description: 'Order detail.', schema: { $ref: getSchemaPath(Order) } })
  getOrder(@Param('id') id: string) {
    return { id, status: 'pending' };
  }

  @Patch(':id/status')
  @ApiOperation({ summary: 'Update order status' })
  @ApiBody({
    schema: {
      type: 'object',
      required: ['status'],
      properties: {
        status: { type: 'string', enum: ['processing', 'shipped', 'delivered', 'cancelled'] },
        reason: { type: 'string' },
        notifyCustomer: { type: 'boolean', default: true },
        trackingNumber: { type: 'string' },
        trackingCarrier: { type: 'string', enum: ['ups', 'fedex', 'usps', 'dhl', 'other'] },
      },
    },
  })
  @ApiOkResponse({ description: 'Order status updated.' })
  updateOrderStatus(@Param('id') id: string, @Body() body: any) {
    return { id, ...body };
  }

  @Delete(':id')
  @ApiOperation({ summary: 'Cancel an order' })
  @ApiNoContentResponse({ description: 'Order cancelled.' })
  cancelOrder(@Param('id') id: string) {
    return;
  }

  @Post(':id/items')
  @ApiOperation({ summary: 'Add an item to an order' })
  @ApiBody({ schema: { $ref: getSchemaPath(OrderLineItem) } })
  @ApiCreatedResponse({ description: 'Item added to order.' })
  addOrderItem(@Param('id') id: string, @Body() body: any) {
    return { orderId: id, item: body };
  }

  @Delete(':id/items/:itemId')
  @ApiOperation({ summary: 'Remove an item from an order' })
  @ApiNoContentResponse({ description: 'Item removed from order.' })
  removeOrderItem(@Param('id') id: string, @Param('itemId') itemId: string) {
    return;
  }

  @Get(':id/tracking')
  @ApiOperation({ summary: 'Get shipment tracking' })
  @ApiOkResponse({
    description: 'Order tracking information.',
    schema: {
      type: 'object',
      properties: {
        trackingNumber: { type: 'string' },
        carrier: { type: 'string' },
        estimatedDelivery: { type: 'string', format: 'date' },
        events: { type: 'array', items: { $ref: getSchemaPath(ShipmentTrackingEventNotificationDelivery) } },
      },
    },
  })
  getTracking(@Param('id') id: string) {
    return { trackingNumber: null, carrier: null, events: [] };
  }

  @Post(':id/refund')
  @ApiOperation({ summary: 'Initiate a refund for an order' })
  @ApiBody({
    schema: {
      type: 'object',
      required: ['reason'],
      properties: {
        amount: { type: 'number', description: 'Partial refund amount. Omit for full refund.' },
        reason: { type: 'string', enum: ['defective', 'not_as_described', 'not_received', 'changed_mind', 'other'] },
        note: { type: 'string', maxLength: 500 },
        items: {
          type: 'array',
          items: {
            type: 'object',
            properties: { itemId: { type: 'string' }, quantity: { type: 'integer' } },
          },
        },
        refundShipping: { type: 'boolean', default: false },
      },
    },
  })
  @ApiCreatedResponse({ description: 'Refund initiated.' })
  createRefund(@Param('id') id: string, @Body() body: any) {
    return { id: 'ref_1', orderId: id, status: 'pending', ...body };
  }

  @Get(':id/invoice')
  @ApiOperation({ summary: 'Get order invoice' })
  @ApiQuery({ name: 'format', required: false, enum: ['json', 'pdf'], schema: { default: 'json' } })
  @ApiOkResponse({ description: 'Order invoice.' })
  getInvoice(@Param('id') id: string) {
    return { orderId: id, invoiceNumber: 'INV-001' };
  }

  @Post(':id/notes')
  @ApiOperation({ summary: 'Add a note to an order' })
  @ApiBody({
    schema: {
      type: 'object',
      required: ['content'],
      properties: {
        content: { type: 'string', maxLength: 1000 },
        isInternal: { type: 'boolean', default: true },
      },
    },
  })
  @ApiCreatedResponse({ description: 'Note added.' })
  addOrderNote(@Param('id') id: string, @Body() body: any) {
    return { id: 'note_1', orderId: id, ...body, createdAt: new Date().toISOString() };
  }

  @Get(':id/timeline')
  @ApiOperation({ summary: 'Get order event timeline' })
  @ApiOkResponse({
    description: 'Order event timeline.',
    schema: {
      type: 'array',
      items: {
        type: 'object',
        properties: {
          id: { type: 'string' },
          type: { type: 'string', enum: ['created', 'status_changed', 'note_added', 'refund_requested', 'shipped'] },
          description: { type: 'string' },
          actor: { type: 'object', properties: { id: { type: 'string' }, name: { type: 'string' }, type: { type: 'string', enum: ['user', 'system'] } } },
          timestamp: { type: 'string', format: 'date-time' },
          metadata: { type: 'object', additionalProperties: true },
        },
      },
    },
  })
  getTimeline(@Param('id') id: string) {
    return [];
  }
}