# apps/example-ui/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bookstore API",
    "description": "Tiny inline OpenAPI document used to demo the React UI component.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://api.bookstore.example.com", "description": "Production" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "Order": {
        "type": "object",
        "required": ["bookId", "quantity"],
        "properties": {
          "bookId": { "type": "string", "format": "uuid" },
          "quantity": { "type": "integer", "minimum": 1 }
        }
      },
...
    }
  },
  "paths": {
...
    "/orders": {
      "post": {
        "operationId": "createOrder",
        "summary": "Place an order",
        "tags": ["orders"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/Order" } }
          }
        },
        "responses": {
          "201": { "description": "Order accepted" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}